refactor: some renames and tweaks
