+++ packages/scaffold/src/core/tester/test-bundler-template/bootsrtap.ts
@@ -1,4 +1,4 @@
-import { TESTER_PLUGIN_ID, TESTER_PLUGIN_REF } from "../../../constant.js";
+import { TESTER_PLUGIN_REF } from "../../../constant.js";
 
 export function generateBootstrap(options: {
   port: number;
@@ -99,21 +99,4 @@ function waitUtilAsync(condition, interval = 100, timeout = 1e4) {
   });
 }
 `;
-}
-
-export function generateManifest() {
-  return {
-    manifest_version: 2,
-    name: "Zotero Plugin Scaffold Test Runner",
-    version: "0.0.1",
-    description: "Test suite for the Zotero plugin. This is a runtime-generated plugin only for testing purposes.",
-    applications: {
-      zotero: {
-        id: TESTER_PLUGIN_ID,
-        update_url: "https://example.com",
-        // strict_min_version: "*.*.*",
-        strict_max_version: "999.*.*",
-      },
-    },
-  };
 }
