+++ packages/scaffold/src/core/tester/test-bundler-template/index.ts
@@ -1,0 +1,4 @@
+export * from "./bootsrtap.js";
+export * from "./html.js";
+export * from "./manifest.js";
+export * from "./mocha-setup.js";
